makefile-improvement (#107)
* auto generate pre-processing targets

* start new makefile

Start rebuilding the makefile from ground up. The old makefile had issues with files from one build interfering with the next and thus required a clean up before each build. The goal for the new makefile is to avoid unnecessary rebuilds of output files by defining proper dependencies and more modular targets.

* add readings rule

* edit comments and remove test rule

* add markdown preprocessing for hugo

* comments

* split targets for web and slides

Motivated by the different requirements of 'web' and 'slides' processing. In both cases images need to be located at the relative positon stated in markdown files. However, the 'web' creates a folder with the same substructure as the source folder but with alterations to original markdown files so that they can be correctly translated by hugo. This folder shouls also not contain any source or intermediate files as those would be included in the final website. The slides on the other hand are generated from the unaltered markdown files. The solution chosen here was to create image files in the same folder as their source files. This way the files are in the right position for slides generation and can be copied to web folder then needed.

* update 'clean' target

* comment

* add target for slides

* fix target for image folders

* add targtes for slides and website creation

* add .noslides marker files

Folders containing a .noslides file will not be converted into slides.

* make slides directly dependend on image files

Defining a phony target that bundels image targets in the same folder and setting it as a prerequisite for the pdf file that uses those images did not work. The pdf file would always be rebuild even if nothing had changed. The pdf files now depend directly on the required image files. Also added an empty target for image files that are not generated. This allows them to be used as prerequisites so that dependend targets get rebuild if the imges are modified/replaced manually.

* cleanup

* rename variable

* comments

* add targets for lecture creation

copied from old makefile

* add target documentation

* comments

* rename variable

* rename variables

* add local.yaml as prerequisite

* add short targets for slides

* modularize docker call

* fixed non-pandoc commands

* add helper for listing slides targets

* more cleaning options

* separate generated files from source dir

Most generated files are now placed outside the source directory. Previously generated images would appear as untracked files in git because the pattern matching for gitignore-files is not powerfull enough to distinguish them from non-generated images. The exception are some intermediate files generated by LaTeX, which still end up in the source directory tree. However, these files can be easily handled by gitignore.

* use canned recipes

* consolidate temporary directories

* update gitignore

* remove unnecessary if-clause

* make canned recipes more backwards compatible

Assignment tokens like '=' and ':=' are not supported before version 3.82.

* no more temp-files in source dir

No temporary files will be generated in the source directory any more. This was achieved by copying all tex-files from the source directory to the temporary directory and have the images generated there. Temporary files from the LaTeX runs are thus contained in the temporary directory like all other temporary files.

* update clean-targets

* Prepare files for new Makefile (#120)

* no slides for misc/resources

* fix outcomes

* fix markdown files

* fix A1

* fix misc

* fix attribgrammars

* fix white space

* escape remaining short codes

* rename mostlyclean to clean

Co-authored-by: Jonas Posselt <jonas.posselt@web.de>